+++ utils/getUpfList.tsx
@@ -1,3 +1,5 @@
+import { HTTPStatus } from "@/utils/utils";
+
 export interface UpfItem {
   hostname: string;
   port: string;
@@ -12,10 +14,10 @@ export const getUpfList = async (token: string): Promise<UpfItem[]> => {
         "Content-Type": "application/json",
       },
     });
-    if (!response.ok) {
-      throw new Error("Failed to fetch UPF list");
-    }
     const upfList = await response.json();
+    if (!response.ok) {
+      throw new Error(`${response.status}: ${HTTPStatus(response.status)}. ${upfList.error}`)
+    }
     return upfList;
   } catch (error) {
     console.error(error);
